refactor(web): migrate simple Jotai atoms to ReScript (migration Phase 2)
Add the reusable Jotai binding (atoms/Jotai.res): `type t<'a>` mapped via
genType onto Jotai's real PrimitiveAtom<T>, plus `external atom`. Every
migrated store is then a genType one-liner, and TS-side
useAtom/useAtomValue/useSetAtom keep inferring the exact value type.

Migrated the low-risk subset (10 modules, ~15 atoms):
- primitives: Drawer, FollowingFeed, FullscreenPlayer, Shortcuts,
  SearchModal (open flags), Feed (feedAtom/feedGeneratorUriAtom)
- string-literal unions -> polymorphic variants (compile to the same JS
  strings, render as the same TS unions): Player.kind, SearchModal.scope
- dicts of primitives -> Dict.t: Feed.feedUrisAtom, Likes
- array of an already-migrated record: Apikeys (array<ApiKey.t>)

Consumers switched their import to the matching `.gen` module; the two
external `SearchScope` importers now take `type scope as SearchScope` from
SearchModal.gen. Local usage (setPlayer("rockbox"), scope compares, dict
spreads) is unchanged.

Deferred on purpose — these migrate with their feature slices later:
- the player/queue cluster (queue.ts's QueueTrack is shared by devices.ts
  and resume.ts) and all large nested-object stores (song, shouts,
  profile(s), stats, top*, recentTracks, lovedTracks, nowpaying, artist,
  dropbox, googledrive, userNowplaying)
- atomWithStorage stores (equalizer, playback, resume) — need a separate,
  faithfully-typed binding (writable + RESET), not the plain `atom` one
- stores holding Set (follows), functions (deviceCommand, playerControls),
  unknown, or react's Key (tab)

Verified: rescript build, tsc -b, and full vite build all pass.
