+++ apps/web/src/atoms/Apikeys.res
@@ -1,0 +1,2 @@
+@genType
+let apiKeysAtom: Jotai.t<array<ApiKey.t>> = Jotai.atom([])
