+++ apps/web/src/atoms/Drawer.res
@@ -1,0 +1,2 @@
+@genType
+let displayDrawerAtom: Jotai.t<bool> = Jotai.atom(false)
