+++ apps/web/src/atoms/FollowingFeed.res
@@ -1,0 +1,2 @@
+@genType
+let followingFeedAtom: Jotai.t<bool> = Jotai.atom(false)
