+++ apps/web/src/atoms/FullscreenPlayer.res
@@ -1,0 +1,2 @@
+@genType
+let fullscreenPlayerAtom: Jotai.t<bool> = Jotai.atom(false)
