+++ apps/web/src/atoms/Likes.res
@@ -1,0 +1,2 @@
+@genType
+let likesAtom: Jotai.t<Dict.t<bool>> = Jotai.atom(Dict.make())
