+++ apps/web/src/atoms/SearchModal.res
@@ -1,27 +1,22 @@
-import { atom } from "jotai";
-
 /** Scope the command palette can be filtered to. */
-export type SearchScope =
-  | "all"
-  | "tracks"
-  | "artists"
-  | "albums"
-  | "playlists"
-  | "users";
+@genType
+type scope = [#all | #tracks | #artists | #albums | #playlists | #users]
 
 /**
  * Whether the Raycast-style command-palette search modal is open. Opened by
  * clicking/focusing the sidebar search input or pressing "/" (see
  * components/KeyboardShortcuts), rendered once globally from the root route.
  */
-export const searchModalOpenAtom = atom<boolean>(false);
+@genType
+let searchModalOpenAtom: Jotai.t<bool> = Jotai.atom(false)
 
 /**
  * Scope the palette should start on when next opened. Set alongside
  * searchModalOpenAtom so shortcuts can open the palette pre-filtered (e.g. "a"
  * → artists); the sidebar trigger and "/" use "all".
  */
-export const searchModalScopeAtom = atom<SearchScope>("all");
+@genType
+let searchModalScopeAtom: Jotai.t<scope> = Jotai.atom(#all)
 
 /**
  * Whether the LIBRARY quick-search palette is open — searches the authenticated
@@ -29,4 +24,5 @@ export const searchModalScopeAtom = atom<SearchScope>("all");
  * actions on results. Opened with Shift+L (see components/KeyboardShortcuts);
  * only meaningful for signed-in users.
  */
-export const librarySearchOpenAtom = atom<boolean>(false);
+@genType
+let librarySearchOpenAtom: Jotai.t<bool> = Jotai.atom(false)
