+++ apps/web/src/atoms/Shortcuts.res
@@ -1,0 +1,3 @@
+/** Whether the keyboard-shortcuts help modal (opened with "?") is visible. */
+@genType
+let shortcutsHelpOpenAtom: Jotai.t<bool> = Jotai.atom(false)
