--- apps/web/src/atoms/apikeys.ts
@@ -1,4 +1,0 @@
-import { atom } from "jotai";
-import type { t as ApiKey } from "../types/ApiKey.gen";
-
-export const apiKeysAtom = atom<ApiKey[]>([]);
