--- apps/web/src/atoms/drawer.ts
@@ -1,3 +1,0 @@
-import { atom } from "jotai";
-
-export const displayDrawerAtom = atom<boolean>(false);
