--- apps/web/src/atoms/followingFeed.ts
@@ -1,3 +1,0 @@
-import { atom } from "jotai";
-
-export const followingFeedAtom = atom<boolean>(false);
