--- apps/web/src/atoms/fullscreenPlayer.ts
@@ -1,3 +1,0 @@
-import { atom } from "jotai";
-
-export const fullscreenPlayerAtom = atom<boolean>(false);
