--- apps/web/src/atoms/likes.ts
@@ -1,5 +1,0 @@
-import { atom } from "jotai";
-
-export const likesAtom = atom<{
-  [key: string]: boolean;
-}>({});
