--- apps/web/src/atoms/shortcuts.ts
@@ -1,4 +1,0 @@
-import { atom } from "jotai";
-
-/** Whether the keyboard-shortcuts help modal (opened with "?") is visible. */
-export const shortcutsHelpOpenAtom = atom<boolean>(false);
