+++ apps/web/src/hooks/useUploadScrobble.ts
@@ -3,7 +3,7 @@ import { useAtomValue } from "jotai";
 import { useEffect, useRef } from "react";
 import { submitScrobble } from "../api/scrobbles";
 import { nowPlayingAtom } from "../atoms/nowpaying";
-import { playerAtom } from "../atoms/player";
+import { playerAtom } from "../atoms/Player.gen";
 import { queueAtom, queueIndexAtom } from "../atoms/queue";
 import { consola } from "consola";
