+++ apps/web/src/pages/home/feed/Feed.tsx
@@ -10,8 +10,8 @@ import relativeTime from "dayjs/plugin/relativeTime";
 import { useAtomValue } from "jotai";
 import { useEffect, useRef } from "react";
 import ContentLoader from "react-content-loader";
-import { feedGeneratorUriAtom } from "../../../atoms/feed";
-import { followingFeedAtom } from "../../../atoms/followingFeed";
+import { feedGeneratorUriAtom } from "../../../atoms/Feed.gen";
+import { followingFeedAtom } from "../../../atoms/FollowingFeed.gen";
 import Handle from "../../../components/Handle";
 import SongCover from "../../../components/SongCover";
 import { WS_URL } from "../../../consts";
